Resolve #80 Add Full Selection Rule to Users
This adds an attribute to users that can be checked to see whether the user is a full allocation partner or a 50% allocation partner. An Integer value of 1 means we will divide the number of open_positons that a partner has by 1. An integer value of 2 means we will divide it by 2 for 50% selection sites. By default sites are 50% allocation sites and the rake task will update users who are 100% allocation sites to be so.