Skip phone number if it is blank
Production was throwing an error on this line when we attempted to import applicants. Presumably this is the result of importing an applicant who has an entry for a mobile or home phone number, but does not have an actual phone number in the phone number field. This will skip those entries and allow the import to continue.