Resolve #10 By Updating Algorithm
This resolves #10 by changing the algorithm to score applicants in relation to jobs by normalizing their distance scores. Instead of scoring each person by actual travel time, we are scoring folks by how relatively close or far they are to the position. So a person whose closest position is 15 minutes away will score the same relative to that position as someone who is 10 minutes away and has it as their closest position. The result of this is farther out applicants will win against closer applicants for a position if they do not have better options.